created 3 images for debugging [rgb]_640x360.png , atlas still shows
even though there is absolutely no loading of it anywhere.
leads me to believe it is indeed framebuffer weirdness.

probably going to say screw it with the framebuffer and
experiment with other render techniques, might someday come back
to it when I'm not strapped for time...
a shame since framebuffers are considered the fastest method and
since other techniques tend to run slower on ios...

will search for a good-enough approach...